feat: Utility to update distributed funds from Spreadsheet ledger file.
