+++ milestone-tracker-importer/data-manager/main.py
@@ -1,8 +1,15 @@
 import typer
 
+from manager.commands import fund9Manager
 from manager.commands import manager
 
 app = typer.Typer()
+
+app.add_typer(
+    fund9Manager.app,
+    name="fund9-manager",
+    help="Manage Fund 9 data to supabase."
+)
 
 app.add_typer(
     manager.app,
