+++ milestone-tracker-importer/data-manager/manager/utils/sb.py
@@ -1,4 +1,5 @@
 from supabase import create_client
+import orjson
 
 from manager.config import settings
 
@@ -21,14 +22,28 @@ class SB():
         results = self.client.table(entity).insert(proposals).execute()
         return results.data
 
-    def upsert_entities(self, proposals, entity):
-        results = self.client.table(entity).upsert(proposals).execute()
+    def upsert_entities(self, _data, entity, fields=None):
+        data = self.prepare_data(_data, fields)
+        results = self.client.table(entity).upsert(data).execute()
         return results.data
 
-    def get_entity(self, entity):
+    def get_entity(self, entity, model=None):
         results = self.client.table(entity).select("*").execute()
-        return [AttrDict(el) for el in results.data]
+        if model:
+            return [model(**el) for el in results.data]
+        else:
+            return [AttrDict(el) for el in results.data]
     
     def create_user(self, email, password):
         user = self.client.auth.sign_up({ "email": email, "password": password })
         return user
+    
+    def prepare_data(self, _data, fields=None):
+        data = []
+        for element in _data:
+            if fields:
+                _serialized = element.json(include=fields)
+            else:
+                _serialized = element.json()
+            data.append(orjson.loads(_serialized))
+        return data
